Refactored IType Added: - Implicit casting - Operator Support

+++ src/BBAP/Lexer/Tokens/Values/IntValueToken.cs
@@ -3,11 +3,11 @@
 namespace BBAP.Lexer.Tokens.Values;
 
 public class IntValueToken : IToken{
-    public IntValueToken(int value, int line) {
+    public IntValueToken(long value, int line) {
         Value = value;
         Line = line;
     }
 
     public int Line { get; }
-    public int Value { get; init; }
+    public long Value { get; init; }
 }
